@@ -1,34 +1,23 @@
 {
   "name": "80lk.github.io",
   "private": true,
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "start": "vue-cli-service serve",
-    "build": "vue-cli-service build"
+    "dev": "vite",
+    "build": "vue-tsc && vite build",
+    "preview": "vite preview"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/80LK/80lk.github.io.git"
-  },
-  "author": "80LK",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/80LK/80lk.github.io/issues"
-  },
-  "homepage": "https://github.com/80LK/80lk.github.io#readme",
   "dependencies": {
-    "@vue/cli-plugin-typescript": "^5.0.4",
-    "@vue/cli-service": "^5.0.4",
-    "typescript": "~4.1.5",
-    "vue": "^2.6.11",
-    "vue-class-component": "^7.2.3",
-    "vue-modal-viewer": "^0.1.7",
-    "vue-property-decorator": "^9.1.2",
-    "vue-router": "^3.5.3",
-    "vue-template-compiler": "^2.6.11",
-    "vuex": "^3.6.2",
-    "vuetify": "^2.6.11"
+    "@fortawesome/fontawesome-free": "^6.4.0",
+    "@mdi/font": "^7.2.96",
+    "vue": "^3.2.47",
+    "vue-router": "^4.2.2"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-vue": "^4.1.0",
+    "typescript": "^5.0.2",
+    "vite": "^4.3.9",
+    "vue-tsc": "^1.4.2"
   }
 }
