@@ -1,20 +1,3 @@
 <template>
-  <v-app>
-    <v-main>
-      <MyHeader />
-    </v-main>
-  </v-app>
+  <router-view></router-view>
 </template>
-
-<script lang="ts">
-import { Component, Vue } from "vue-property-decorator";
-import Header from "./components/header.vue";
-
-@Component({
-  components: { MyHeader: Header },
-  data() {
-    return {};
-  },
-})
-export default class App extends Vue {}
-</script>
